Naively updating counter based on packets.
This takes 2 terminal input events to actually show the network-driven update to the counter. `crossterm::event::read` blocks until a terminal input event and we do that *after* receiving any packets. So, if we receive a packet while blocked in `crossterm::event::read`, we stay blocked until a terminal input event, refresh the screen, read the packet and update the counter, and wait for a second input event before refreshing the screen with that update from the packet.